extracted LegacyBsNotification component
